feat: Add routesEnum.js for defining routes
Add a new file, routesEnum.js, to define the routes for the application. This file includes constants for the home, about, projects, works, and contact routes.

refactor: Update AboutView.vue and HomeView.vue

In AboutView.vue, update the layout and content to include an image, information about the developer, and styling changes. In HomeView.vue, update the layout and content to include information about the developer, social links, and the main tech stack.

refactor: Update ContactView.vue, ProjectsView.vue, and WorksView.vue

In ContactView.vue, update the layout and content to include a contact form and location information. In ProjectsView.vue, update the content to indicate that new projects are coming soon. In WorksView.vue, update the content to display the text "Works".